[update]   - Improve Fortran/C string-length handling in `gtool_file_f.c` for better Mac/Linux compatibility:   - unify hidden Fortran character-length arguments through a configurable `fortran_charlen_t`,   - default to `size_t` and expose the choice via `GTOOL_FC_CHARLEN_T`,   - harden `fstr2cstr` for zero-length and all-space strings to avoid invalid C-string termination. - Propagate the string-length ABI choice to GNU/Intel `sysdep/Makedef.*` via explicit `CFLAGS_SYSDEPEND` definitions for consistent builds. - Increase `History_axis_limit` to reduce axis-definition failures in tracking-heavy outputs. - Align SDM motion/coalescence interfaces and numeric handling:   - promote `sd_ncol` to a wider integer kind for large coalescence counts,   - switch selected motion arguments to `intent(inout)` to match in-place update behavior,   - replace hard-coded `1.d0` with `1.0_RP` for kind consistency. - Refine SDM I/O behavior:   - import and use forward/backward tracking flags consistently,   - normalize NetCDF deflate/shuffle flags to integer values,   - add a safe fallback assignment before abort paths. - Update case namelists to avoid invalid double-precision literal forms while preserving valid entries such as `86.7132D0`. - Add acknowledgements to `README.md`.

+++ contrib/SDM/sdm_coalescence_cold.f90
@@ -139,7 +139,7 @@ contains
     real(RP), intent(out) :: c_rate(1:sd_num) ! coalescence probability
     ! Internal shared variables
     real(RP) :: sd_aslrho(1:22) ! Density of chemical material contained as water-soluble aerosol in super droplets
-    integer  :: sd_ncol ! how many times coalescence occurs
+    integer(DP)  :: sd_ncol ! how many times coalescence occurs
     integer :: freq_max ! get the maximum number of super-droplets in each grid
     integer :: hfreq_max ! hfreq_max / 2
     integer :: ipremium  ! premium coef. for coalescence
@@ -1325,7 +1325,7 @@ contains
     ! calculate the outcome of droplet-droplet coalescence
     use m_sdm_common, only: O_THRD
 
-    integer, intent(in)     :: sd_ncol ! how many times coalescence occurs
+    integer(DP), intent(in)     :: sd_ncol ! how many times coalescence occurs
     real(RP), intent(in)    :: sd_r1   ! radius of super-droplets with larger multiplicity
     real(RP), intent(inout) :: sd_r2   ! radius of super-droplets with smaller multiplicity
 
@@ -1352,7 +1352,7 @@ contains
 
     use m_sdm_common, only: O_THRD,F_THRD,ONE_PI,STAT_ICE,STAT_LIQ,i2
 
-    integer, intent(in)     :: sd_ncol ! how many times coalescence occurs
+    integer(DP), intent(in)     :: sd_ncol ! how many times coalescence occurs
     real(RP), intent(in)    :: sd_r1   ! radius of liquid droplet with larger multiplicity
     real(RP), intent(in)    :: sd_r2   ! radius of liquid droplet with smaller multiplicity
     real(RP), intent(in)    :: sd_re1  ! equatorial radius of ice particle with larger multiplicity
@@ -1730,7 +1730,7 @@ contains
          rhoi_mks => CONST_DICE  ! density of ice [kg/m^3]
     use m_sdm_common, only: F_THRD,ONE_PI
 
-    integer, intent(in)     :: sd_ncol ! how many times coalescence occurs
+    integer(DP), intent(in)     :: sd_ncol ! how many times coalescence occurs
     real(RP), intent(in)    :: sd_re1  ! equatorial radius of ice particle with larger multiplicity
     real(RP), intent(inout) :: sd_re2  ! equatorial radius of ice particle with smaller multiplicity
     real(RP), intent(in)    :: sd_rp1  ! polar radius of ice particle with larger multiplicity
@@ -1800,7 +1800,7 @@ contains
          rhoi_mks => CONST_DICE  ! density of ice [kg/m^3]
     use m_sdm_common, only: F_THRD,ONE_PI
 
-    integer, intent(in)     :: sd_ncol ! how many times coalescence occurs
+    integer(DP), intent(in)     :: sd_ncol ! how many times coalescence occurs
     real(RP), intent(in)    :: sd_re1  ! equatorial radius of ice particle with larger multiplicity
     real(RP), intent(inout) :: sd_re2  ! equatorial radius of ice particle with smaller multiplicity
     real(RP), intent(in)    :: sd_rp1  ! polar radius of ice particle with larger multiplicity
